requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
pypdf
python-dotenv
requests
tiktoken